Readme corrections
Adjusted mentions of how some O2s use ARCS for booting, plus tweaked the wording to avoid confusing the O2 as an R4xxx system since it isn't. Updated some of the installation procedure for OS X systems to state that installing Ansible through brew installs python as a dependency, and included an alternative installation procedure if the user already installed python outside of brew.